+++ preprocess.h
@@ -4,10 +4,10 @@
 extern "C" {
     void preprocess(
 	        struct SW_Data_Packet * input_data_packet, // Read-Only Data Packet Struct
-	        uint16_t *input_all_peds, // Read-Only Pedestals
+	        unsigned *input_all_peds, // Read-Only Pedestals
             int * bounds, // Read-Only Integral Bounds
-            int32_t *zero_thresholds, // Read-Only Thresholds for zero-suppression
-	        int32_t *output_integrals,       // Output Result (Integrals)
+            int *zero_thresholds, // Read-Only Thresholds for zero-suppression
+	        int *output_integrals,       // Output Result (Integrals)
             struct Centroid *centroid // Output Centroid
 	);
 }
